backend/terrain_splitter/pyproject.toml
[project]
name = "terrain-splitter"
version = "0.1.0"
description = "Local backend terrain partition solver for MapPolyAngle"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.34,<1.0",
  "mangum>=0.19,<1.0",
  "numpy>=2.1,<3.0",
  "pillow>=11.1,<12.0",
  "shapely>=2.1,<3.0",
  "httpx>=0.28,<1.0",
  "pyproj>=3.7,<4.0",
  "tifffile>=2025.6,<2026.0",
  "imagecodecs>=2026.1,<2027.0",
  "python-multipart>=0.0.20,<1.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<1.0",
  "pre-commit>=4.2,<5.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["terrain_splitter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
  ".aws-sam",
  ".cache",
  ".debug",
  ".venv",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
